requirements.txt
loguru==0.7.3
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
PyYAML==6.0.2
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
torch==2.6.0

